use path aliases for cleaner imports; separate item queries to dedicated file
